+++ lib/engine.rb
@@ -1,4 +1,3 @@
-require 'pry'
 # Defining the engine class.
 class Engine
   def initilize; end